bevy_asset_loader::asset_collection
Trait AssetCollectionApp

pub trait AssetCollectionApp {
    // Required method
    fn init_collection<A: AssetCollection>(&mut self) -> &mut Self;
}

Extension trait for App enabling initialisation of asset collections

Required Methods

fn init_collection<A: AssetCollection>(&mut self) -> &mut Self

Initialise an AssetCollection

This function does not give any guaranties about the loading status of the asset handles. If you want to use a loading state, you do not need this function! Instead use an LoadingState and add collections to it to be prepared during the loading state.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.