bevy_asset_loader::asset_collection
Trait AssetCollectionWorld

pub trait AssetCollectionWorld {
    // Required method
    fn init_collection<A: AssetCollection>(&mut self);
}

Extension trait for World enabling initialisation of asset collections

Required Methods

fn init_collection<A: AssetCollection>(&mut self)

Initialise an AssetCollection

This function does not give any guaranties about the loading status of the asset handles. If you want such guaranties, use a LoadingState.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.