Module bevy_asset_loader::dynamic_asset

Types and infrastructure to load and use dynamic assets

Structs

DynamicAssetCollections

Resource keeping track of dynamic asset collection files for different loading states

DynamicAssets

Resource to dynamically resolve keys to assets.

Enums

DynamicAssetType

Different typed that can generate the asset field value of a dynamic asset

Traits

DynamicAsset

Any type implementing this trait can be assigned to asset keys as part of a dynamic asset collection.

DynamicAssetCollection

This traits describes types that contain asset configurations and can register them in the DynamicAssets resource.