bevy_asset_loader
Module loading_state

A game state responsible for loading assets

Modules

config

Configuration of loading states

Structs

LoadingState

A Bevy plugin to configure automatic asset loading

LoadingStateSchedules

Resource to store the schedules for loading states

LoadingStateSet

Systems in this set check the loading state of assets.

Traits

LoadingStateAppExt

Extension trait for Bevy Apps to add loading states idiomatically