Trait bevy_asset_loader::loading_state::LoadingStateAppExt

pub trait LoadingStateAppExt {
    fn add_loading_state<S: States>(
        &mut self,
        loading_state: LoadingState<S>
    ) -> &mut Self;
    fn add_collection_to_loading_state<S: States, A: AssetCollection>(
        &mut self,
        loading_state: S
    ) -> &mut Self;
    fn add_dynamic_collection_to_loading_state<S: States, C: DynamicAssetCollection + Asset>(
        &mut self,
        loading_state: S,
        file: &str
    ) -> &mut Self;
    fn init_resource_after_loading_state<S: States, A: Resource + FromWorld>(
        &mut self,
        loading_state: S
    ) -> &mut Self;
}

Extension trait for Bevy Apps to add loading states idiomatically

Required Methods

fn add_loading_state<S: States>(
&mut self,
loading_state: LoadingState<S>
) -> &mut Self

Add a loading state to your app

fn add_collection_to_loading_state<S: States, A: AssetCollection>(
&mut self,
loading_state: S
) -> &mut Self

Add an AssetCollection to the LoadingState

The added collection will be loaded and inserted into your Bevy app as a resource.

    App::new()
        .add_loading_state(
          LoadingState::new(GameState::Loading)
            .continue_to_state(GameState::Menu)
        )
        .add_collection_to_loading_state::<_, AudioAssets>(GameState::Loading)
        .add_collection_to_loading_state::<_, ImageAssets>(GameState::Loading)

fn add_dynamic_collection_to_loading_state<S: States, C: DynamicAssetCollection + Asset>(
&mut self,
loading_state: S,
file: &str
) -> &mut Self

Register files to be loaded as a certain type of DynamicAssetCollection

During the loading state, the given dynamic asset collections will be loaded and their content registered. This will happen before trying to resolve any dynamic assets as part of asset collections.

You need to register a loader for your asset type yourself. If you want to see some code, take a look at the custom_dynamic_assets example.

fn init_resource_after_loading_state<S: States, A: Resource + FromWorld>(
&mut self,
loading_state: S
) -> &mut Self

Add any FromWorld resource to be initialized after all asset collections are loaded.

    App::new()
        .add_loading_state(
          LoadingState::new(GameState::Loading)
            .continue_to_state(GameState::Menu)
        )
        .add_collection_to_loading_state::<_, TextureForAtlas>(GameState::Loading)
        .init_resource_after_loading_state::<_, TextureAtlasFromWorld>(GameState::Loading)